Extract volume-specific script
This needs parameterisation but I don’t want to change
everything at once.
